pyproject.toml
[project]
name = "web-search-mcp"
version = "0.1.0"
description = "MCP server that allows to search web using Kagi API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "mcp>=1.1.1",
 "python-dotenv>=1.0.1",
]
[[project.authors]]
name = "Aleksey Pridachin"
email = "a.pridachin@semrush.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
web-search-mcp = "web_search_mcp:main"
